Resolve test case
