+++ .scrutinizer.yml
@@ -1,0 +1,32 @@
+filter:
+    excluded_paths: [tests/*]
+
+checks:
+    php:
+        code_rating: true
+        duplication: true
+tools:
+    php_analyzer: true
+    php_mess_detector: true
+    php_changetracking: true
+    php_code_sniffer:
+        config:
+            standard: PSR2
+    php_loc:
+        excluded_dirs:
+            - vendor
+    php_pdepend:
+        excluded_dirs:
+            - vendor
+            - tests
+build:
+    environment:
+        php:
+            version: 7.2       # Common versions: 5.4, 5.5, 5.6, 7.0, 7.1, or hhvm
+    tests:
+        override:
+            -
+                command: 'vendor/bin/phpunit --coverage-clover=some-file'
+                coverage:
+                    file: 'some-file'
+                    format: 'clover'
