+++ tests/Stubs/StubAuthorizableClass.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Author       : Samuel Dervis
+ * Email        : sam@samerior.com
+ * Company      : Samerior Group
+ */
+
+namespace Samerior\LaravelSidebar\Tests\Stubs;
+
+use Samerior\LaravelSidebar\Traits\AuthorizableTrait;
+
+/**
+ * Class StubAuthorizableClass
+ * @package Samerior\LaravelSidebar\Tests\Stubs
+ */
+class StubAuthorizableClass
+{
+    use AuthorizableTrait;
+}
